fix: adapt Runner and ResourceCompiler to changes in Compiler

+++ src/main/java/compiler/ResourceCompiler.java
@@ -16,7 +16,7 @@ public class ResourceCompiler {
 
 			if (isSourceFile(fileName)) {
 
-				int[] bytecode = compiler.compile(fileName);
+				byte[] bytecode = compiler.compile(fileName);
 			}
 		}
